# server/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "REST API",
    "description": "RestAPI endpoints"
  },
  "host": "localhost:5000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "definitions": {
    "User": {
      "type": "object",
      "properties": {
        "login": { "type": "string" },
        "password": { "type": "string" },
        "firstName": { "type": "string" },
        "lastName": { "type": "string" },
        "email": { "type": "string" },
        "phone": { "type": "string" },
        "address": { "type": "string" },
        "gender": { "type": "string", "enum": ["male", "female"] },
        "birthDate": { "type": "string", "format": "date-time" },
        "profilePictureURL": { "type": "string" },
        "education": { "type": "string" },
        "occupation": { "type": "string" },
        "bio": { "type": "string" },
        "nationality": { "type": "string" },
        "relationshipStatus": { "type": "string", "enum": ["single", "married", "divorced"] },
        "accountCreationDate": { "type": "string", "format": "date-time" },
        "lastLogin": { "type": "string", "format": "date-time" },
        "isActive": { "type": "boolean" },
        "interests": { "type": "array", "items": { "type": "string" } },
        "skills": { "type": "array", "items": { "type": "string" } },
        "socialMediaLinks": { "type": "array", "items": { "type": "string" } },
        "friends": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/User"
          }
        }
      }
    },
    "Post": {
      "type": "object",
      "properties": {
        "date": { "type": "string", "format": "date-time" },
        "category": { "type": "string" },
        "content": { "type": "string" },
        "likeCount": { "type": "integer" },
        "authorId": { "type": "string","format": "ObjectId" }
      }
    }
  },
  "paths": {
    "/api/users": {
      "get": {
        "description": "Get all users",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": { "$ref": "#/definitions/User" }
            }
          }
        }
      },
      "post": {
        "description": "Create a new user",
        "responses": {
          "200": {
            "description": "OK",
            "schema": { "$ref": "#/definitions/User" }
          }
        }
      }
    },
    "/api/users/{id}": {
      "get": {
        "description": "Get a user by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": { "$ref": "#/definitions/User" }
          }
        }
      },
      "put": {
        "description": "Update a user by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": { "$ref": "#/definitions/User" }
          }
        }
      },
      "delete": {
        "description": "Delete a user by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/posts": {
      "get": {
        "description": "Get all posts",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": { "$ref": "#/definitions/Post" }
            }
          }
        }
      },
      "post": {
        "description": "Add a new post",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": { "$ref": "#/definitions/Post" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": { "$ref": "#/definitions/Post" }
          }
        }
      }
    },
    "/api/posts/{id}": {
      "get": {
        "description": "Get a single post by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": { "$ref": "#/definitions/Post" }
          }
        }
      },
      "put": {
        "description": "Update a post by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": { "$ref": "#/definitions/Post" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": { "$ref": "#/definitions/Post" }
          }
        }
      },
      "delete": {
        "description": "Delete a post by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/users/{id}/is-active": {
      "get": {
        "description": "Check if a user is active by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/users/{id}/basic-informations": {
      "get": {
        "description": "Get basic information about a user by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": { "$ref": "#/definitions/User" }
          }
        }
      }
    },
    "/api/users/{id}/posts": {
      "get": {
        "description": "Get posts by a user ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": { "$ref": "#/definitions/Post" }
            }
          }
        }
      }
    },
    "/api/users/{id}/friends": {
      "get": {
        "description": "Get friends of a user by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": { "$ref": "#/definitions/User" }
            }
          }
        }
      }
    },
     "/api/users/{id}/under-fetching-solution": {
    "get": {
      "description": "Get user details, posts, and friends by user ID",
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "type": "string"
        }
      ],
      "responses": {
        "200": {
          "description": "OK",
          "schema": {
            "type": "object",
            "properties": {
              "user": { "$ref": "#/definitions/User" },
              "posts": {
                "type": "array",
                "items": { "$ref": "#/definitions/Post" }
              },
              "friends": {
                "type": "array",
                "items": { "$ref": "#/definitions/User" }
              }
            }
          }
        }
      }
    }
  }
  }
}